src/requirements.txt
boto3==1.10.22
botocore==1.13.22
certifi==2019.9.11
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.5
django-heroku==0.3.1
django-mysql==3.2.0
django-storages==1.7.2
docutils==0.15.2
gunicorn==19.9.0
idna==2.8
jmespath==0.9.4
mysqlclient==1.4.4
numpy==1.17.4
Pillow==6.2.0
psycopg2==2.8.3
python-dateutil==2.8.0
pytz==2019.2
reportlab==3.5.32
requests==2.22.0
s3transfer==0.2.1
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
whitenoise==4.1.4